Restruce tests into folders for logical separation
