+++ EssentialFeed/EssentialFeediOSTests/Feed UI/snapshots/FEED_WITH_ERROR_MESSAGE_light.png
