+++ EssentialFeed/EssentialFeediOSTests/Shared UI/ListSnapshotTests.swift
@@ -1,0 +1,32 @@
+import XCTest
+import EssentialFeediOS
+@testable import EssentialFeed
+
+class ListSnapshotTests: XCTestCase {
+
+    func test_emptyList() {
+        let sut = makeSUT()
+
+        sut.display(emptyList())
+
+        assert(snapshot: sut.snapshot(for: .iPhone8(style: .light)), named: "EMPTY_LIST_light")
+        assert(snapshot: sut.snapshot(for: .iPhone8(style: .dark)), named: "EMPTY_LIST_dark")
+    }
+
+    // MARK: - Helpers
+
+    private func makeSUT() -> ListViewController {
+        let bundle = Bundle(for: ListViewController.self)
+        let storyboard = UIStoryboard(name: "Feed", bundle: bundle)
+        let controller = storyboard.instantiateInitialViewController() as! ListViewController
+        controller.loadViewIfNeeded()
+        controller.tableView.showsVerticalScrollIndicator = false
+        controller.tableView.showsHorizontalScrollIndicator = false
+        return controller
+    }
+
+    private func emptyList() -> [CellController] {
+        return []
+    }
+
+}
