+++ EssentialFeed/EssentialFeediOSTests/Shared UI/snapshots/EMPTY_LIST_light.png
